refactor(crane_session_coordinator): 一度しか使われないクラスを使用箇所へ統合
## 背景
crane_session_coordinator は8クラス（hpp+cpp各ファイル）に分割されていたが、
多くが「1箇所からしか使われない薄いラッパー」になっており、処理の流れを追うために
複数ファイルを行き来する必要があり見通しが悪かった。
一度しか使わないコードは使用箇所の近くに置くことで可読性を向上させる。

## 変更内容
- DiagnosticsReporter を SessionCoordinatorComponent にインライン化
  - recordCycle/updateDiagnosticsの実装を直接展開
  - std::cerr の代わりに RCLCPP_WARN を使用するよう改善
- CommandAggregator を SessionCoordinatorComponent にインライン化
  - collectCommands/setMessageMetadata/assignPriorities を1メソッドに統合
- GlobalRobotAllocator を RobotAllocator にインライン化
  - allocate() を allocateRobotsGreedy() としてインライン化
  - SessionRequirement 構造体を robot_allocator.hpp に移動
- SessionRegistry の4つのワンライナーメソッドをヘッダにインライン定義
  - getOrCreatePlanner のみ cpp に残す

## 結果
ヘッダ 8→6、ソース 8→5（削除ファイル6つ）
